Split out Alfresco, Blockchain Proof and Factom related classes
