+++ src/main/java/com/sphereon/alfresco/blockchain/agent/backend/tasks/scheduled/VerifyPendingRegistrationsTask.java
@@ -2,7 +2,7 @@ package com.sphereon.alfresco.blockchain.agent.backend.tasks.scheduled;
 
 import com.alfresco.apis.handler.ApiException;
 import com.alfresco.apis.model.ResultNode;
-import com.sphereon.alfresco.blockchain.agent.backend.commands.certficate.Signer;
+import com.sphereon.alfresco.blockchain.agent.sphereon.utils.Signer;
 import com.sphereon.alfresco.blockchain.agent.backend.tasks.BlockchainTask;
 import com.sphereon.alfresco.blockchain.agent.backend.tasks.Task;
 import com.sphereon.alfresco.blockchain.agent.sphereon.proof.ProofApiUtils;
@@ -16,7 +16,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import static com.sphereon.alfresco.blockchain.agent.backend.AlfrescoBlockchainRegistrationState.PENDING_VERIFICATION;
+import static com.sphereon.alfresco.blockchain.agent.sphereon.alfresco.AlfrescoBlockchainRegistrationState.PENDING_VERIFICATION;
 
 @Component
 public class VerifyPendingRegistrationsTask implements Task<Void> {
