+++ src/main/java/com/sphereon/alfresco/blockchain/agent/sphereon/proof/CheckBlockchainProofConfigTask.java
@@ -1,11 +1,10 @@
-package com.sphereon.alfresco.blockchain.agent.backend.tasks.ondemand;
+package com.sphereon.alfresco.blockchain.agent.sphereon.proof;
 
+import com.sphereon.alfresco.blockchain.agent.backend.tasks.Task;
 import com.sphereon.sdk.blockchain.proof.api.ConfigurationApi;
 import com.sphereon.sdk.blockchain.proof.handler.ApiException;
 import com.sphereon.sdk.blockchain.proof.model.ChainSettings;
 import com.sphereon.sdk.blockchain.proof.model.CreateConfigurationRequest;
-import com.sphereon.alfresco.blockchain.agent.backend.tasks.Task;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
@@ -18,22 +17,23 @@ import static com.sphereon.sdk.blockchain.proof.model.ChainSettings.HashAlgorith
 import static com.sphereon.sdk.blockchain.proof.model.CreateConfigurationRequest.AccessModeEnum;
 
 @Component
-public class CheckBlockchainConfigTask implements Task<Void> {
-
+public class CheckBlockchainProofConfigTask implements Task<Void> {
     private static final String EXCEPTION_MESSAGE_GET_CONFIG = "An error occurred whilst retrieving blockchain configuration %s: %s";
     private static final String ASSERT_MESSAGE_GET_CONFIG = "An error occurred whilst retrieving blockchain configuration %s. The response was empty.";
     private static final String EXCEPTION_MESSAGE_CREATE_CONFIG = "An error occurred whilst creating blockchain configuration %s: %s";
     private static final String EXCEPTION_MESSAGE_CREATE_CONFIG_API = "An error occurred whilst creating blockchain configuration %s, HTTP status: %d\n Response body:\n%s";
 
-    @Autowired
-    private ConfigurationApi configurationApi;
+    private final ConfigurationApi configurationApi;
+    private final String configName;
+    private final String context;
 
-    @Value("${blockchain.config-name:#{null}}")
-    protected String configName;
-
-    @Value("${blockchain.context:factom}")
-    protected String context;
-
+    public CheckBlockchainProofConfigTask(final ConfigurationApi configurationApi,
+                                          @Value("${blockchain.config-name:#{null}}") final String configName,
+                                          @Value("${blockchain.context:factom}") final String context) {
+        this.configurationApi = configurationApi;
+        this.configName = configName;
+        this.context = context;
+    }
 
     @Override
     public Void execute() {
@@ -42,37 +42,35 @@ public class CheckBlockchainConfigTask implements Task<Void> {
             Assert.notNull(configResponse, String.format(ASSERT_MESSAGE_GET_CONFIG, configName));
             Assert.notNull(configResponse.getConfiguration(), String.format(ASSERT_MESSAGE_GET_CONFIG, configName));
         } catch (ApiException e) {
-            if (e.getCode() == HttpStatus.NOT_FOUND.value()) {
-                createConfiguration();
-            } else {
+            if (e.getCode() != HttpStatus.NOT_FOUND.value()) {
                 throw new RuntimeException(String.format(EXCEPTION_MESSAGE_GET_CONFIG, configName, "" + e.getCode()), e);
             }
-
-        } catch (Throwable throwable) {
-            throw new RuntimeException(String.format(EXCEPTION_MESSAGE_GET_CONFIG, configName, throwable.getMessage()), throwable);
+            createConfiguration();
+        } catch (Exception exception) {
+            throw new RuntimeException(String.format(EXCEPTION_MESSAGE_GET_CONFIG, configName, exception.getMessage()), exception);
         }
         return null;
     }
 
-
     private void createConfiguration() {
         try {
+            final var proofTypes = List.of(ContentRegistrationChainTypesEnum.PER_HASH_PROOF_CHAIN, ContentRegistrationChainTypesEnum.SINGLE_PROOF_CHAIN);
             final var chainSettings = new ChainSettings()
-                .hashAlgorithm(HashAlgorithmEnum._256)
-                .contentRegistrationChainTypes(List.of(ContentRegistrationChainTypesEnum.PER_HASH_PROOF_CHAIN, ContentRegistrationChainTypesEnum.SINGLE_PROOF_CHAIN));
+                    .hashAlgorithm(HashAlgorithmEnum._256)
+                    .contentRegistrationChainTypes(proofTypes);
 
             final var createConfigRequest = new CreateConfigurationRequest()
-                .name(configName)
-                .context(context)
-                .accessMode(AccessModeEnum.PRIVATE)
-                .initialSettings(chainSettings);
+                    .name(configName)
+                    .context(context)
+                    .accessMode(AccessModeEnum.PRIVATE)
+                    .initialSettings(chainSettings);
             final var configResponse = configurationApi.createConfiguration(createConfigRequest);
             Assert.notNull(configResponse, String.format(ASSERT_MESSAGE_GET_CONFIG, configName));
             Assert.notNull(configResponse.getConfiguration(), String.format(ASSERT_MESSAGE_GET_CONFIG, configName));
         } catch (ApiException e) {
             throw new RuntimeException(String.format(EXCEPTION_MESSAGE_CREATE_CONFIG_API, configName, e.getCode(), e.getResponseBody()), e);
-        } catch (Throwable throwable) {
-            throw new RuntimeException(String.format(EXCEPTION_MESSAGE_CREATE_CONFIG, configName, throwable.getMessage()), throwable);
+        } catch (Exception exception) {
+            throw new RuntimeException(String.format(EXCEPTION_MESSAGE_CREATE_CONFIG, configName, exception.getMessage()), exception);
         }
     }
 }
