+++ src/main/java/com/sphereon/alfresco/blockchain/agent/sphereon/proof/config/TokenUpdater.java
@@ -1,4 +1,4 @@
-package com.sphereon.alfresco.blockchain.agent.config;
+package com.sphereon.alfresco.blockchain.agent.sphereon.proof.config;
 
 import com.sphereon.libs.authentication.api.AuthenticationApi;
 import org.springframework.stereotype.Component;
